+++ sporco/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-__version__ = '0.2.2a1'
+__version__ = '0.2.2a2'
 
 
 # This is a temporary solution to the circular imports resulting from the