+++ sporco/pgm/bpdn.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2018-2020 by Brendt Wohlberg <brendt@ieee.org>
+# Copyright (C) 2018-2025 by Brendt Wohlberg <brendt@ieee.org>
 #                            Cristina Garcia-Cardona <cgarciac@lanl.gov>
 # All rights reserved. BSD 3-clause License.
 # This file is part of the SPORCO package. Details of the copyright
@@ -153,7 +153,7 @@ class BPDN(pgm.PGM):
         self.S = np.asarray(S, dtype=self.dtype)
 
         self.Y = self.X.copy()
-        self.Yprv = self.Y.copy() + 1e5
+        self.Yprv = self.Y.copy() + 1e3
 
         self.setdict(D)